Extract ErrorHelpers from CoreComponent
So that `mix phx.gen.json` work with `mix phx.new --no-html` app
because ErrorHelpers is always generated.
