+++ installer/lib/phx_new/single.ex
@@ -13,6 +13,7 @@ defmodule Phx.New.Single do
      "phx_single/lib/app_name/application.ex": "lib/:app/application.ex",
      "phx_single/lib/app_name.ex": "lib/:app.ex",
      "phx_web/controllers/error_json.ex": "lib/:lib_web_name/controllers/error_json.ex",
+     "phx_web/controllers/error_helpers.ex": "lib/:lib_web_name/controllers/error_helpers.ex",
      "phx_web/endpoint.ex": "lib/:lib_web_name/endpoint.ex",
      "phx_web/router.ex": "lib/:lib_web_name/router.ex",
      "phx_web/telemetry.ex": "lib/:lib_web_name/telemetry.ex",
