+++ priv/templates/phx.gen.json/changeset_json.ex
@@ -3,10 +3,10 @@ defmodule <%= inspect context.web_module %>.ChangesetJSON do
   Traverses and translates changeset errors.
 
   See `Ecto.Changeset.traverse_errors/2` and
-  `<%= inspect context.web_module %>.CoreCompponents.translate_error/1` for more details.
+  `<%= inspect context.web_module %>.ErrorHelpers.translate_error/1` for more details.
   """
   def translate_errors(changeset) do
-    Ecto.Changeset.traverse_errors(changeset, &<%= inspect context.web_module %>.CoreComponents.translate_error/1)
+    Ecto.Changeset.traverse_errors(changeset, &<%= inspect context.web_module %>.ErrorHelpers.translate_error/1)
   end
 
   @doc """
